refactor: - calculate parsedRecurrence inside RecurrenceEditor
