+++ src/ui/RecurrenceEditor.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
     import { TASK_FORMATS } from '../Config/Settings';
+    import { Recurrence } from '../Task/Recurrence';
     import type { EditableTask } from './EditableTask';
     import { labelContentWithAccessKey } from './EditTaskHelpers';
 
@@ -8,6 +9,31 @@
     export let accesskey: string | null;
 
     export let parsedRecurrence: string;
+
+    // NEW_TASK_FIELD_EDIT_REQUIRED
+    $: {
+        isRecurrenceValid = true;
+        if (!editableTask.recurrenceRule) {
+            parsedRecurrence = '<i>not recurring</>';
+        } else {
+            const recurrenceFromText = Recurrence.fromText({
+                recurrenceRuleText: editableTask.recurrenceRule,
+                // Only for representation in the modal, no dates required.
+                startDate: null,
+                scheduledDate: null,
+                dueDate: null,
+            })?.toText();
+            if (!recurrenceFromText) {
+                parsedRecurrence = '<i>invalid recurrence rule</i>';
+                isRecurrenceValid = false;
+            } else if (!editableTask.startDate && !editableTask.scheduledDate && !editableTask.dueDate) {
+                parsedRecurrence = '<i>due, scheduled or start date required</i>';
+                isRecurrenceValid = false;
+            } else {
+                parsedRecurrence = recurrenceFromText;
+            }
+        }
+    }
 
     const { recurrenceSymbol } = TASK_FORMATS.tasksPluginEmoji.taskSerializer.symbols;
 </script>
